Switch prices to integer ticks + random-walk mid for benchmark
- double price → int32_t price in Order, Trade, and all maps
  Integer comparison is faster and avoids floating-point rounding in price-level lookups
- Add generate_mid_path(): symmetric ±1-tick random walk clamped to [9000,11000]
  Mid drifts into resting orders so the benchmark exercises matching, not just insertion
- Add generate_orders() using N(mid, σ=5 ticks) per order
  Orders cluster tightly around mid — realistic spread, high cross rate
- Update operator<< to display tick prices as "$D.CC (T ticks)"
- Update general_test() with tick prices (10000=$100.00, etc.)

+++ Order.cpp
@@ -11,13 +11,13 @@ Order::Order()
         : order_id(0),
           side(OrderSide::Buy),
           type(OrderType::Limit),
-          price(0.0),
+          price(0),
           quantity(0),
           timestamp(std::chrono::high_resolution_clock::now())
 {}
 
 // parameterized constructor
-Order::Order(OrderSide s, OrderType t, double p, uint64_t q)
+Order::Order(OrderSide s, OrderType t, int32_t p, uint64_t q)
         : order_id(0),
           client_order_id(0),
           side(s),
